requirements.txt
google-genai>=1.0.0
mplfinance>=0.12.10
python-binance>=1.0.19
qdrant-client>=1.12.0
pandas>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
